Move APIError and ServerError to haas.errors
This is to prevent a circular dependency in a change to be made soon:
we're about to make the database session a request local variable, which
requires importing haas.model from haas.rest. This would introduce the
cycle haas.rest -> haas.model -> haas.errors -> haas.rest. By moving the
above exceptions to haas.errors, it no longer needs to import haas.rest,
breaking the cycle.
