+++ haas/rest.py
@@ -15,8 +15,7 @@
 
 The function `wsgi_handler` is the wsgi entry point to the app.
 
-The decorator `rest_call` and the classes `APIError` and `ServerError`
-are the main things of interest in this module.
+The decorator `rest_call` is the main thing of interest in this module.
 """
 import logging
 import inspect
@@ -27,6 +26,8 @@ from werkzeug.routing import Map, Rule, parse_rule
 from werkzeug.exceptions import HTTPException, InternalServerError
 from werkzeug.local import Local, LocalManager
 
+from haas.errors import APIError, ServerError
+
 from schema import Schema, SchemaError
 
 local = Local()
@@ -35,34 +36,6 @@ local_manager = LocalManager([local])
 logger = logging.getLogger(__name__)
 
 _url_map = Map()
-
-
-class APIError(Exception):
-    """An exception indicating an error that should be reported to the user.
-
-    i.e. If such an error occurs in a rest API call, it should be reported as
-    part of the HTTP response.
-    """
-    status_code = 400  # Bad Request
-
-    def response(self):
-        # TODO: We're getting deprecation errors about the use of self.message.
-        # We should figure out what the right way to do this is.
-        return Response(json.dumps({'type': self.__class__.__name__,
-                                    'msg': self.message,
-                                    }), status=self.status_code)
-
-
-class ServerError(Exception):
-    """An error occurred when trying to process the request.
-
-    This is likely not the client's fault; as such the HTTP status is 500.
-    The semantics are much the same as the corresponding HTTP error.
-
-    In general, we do *not* want to report the details to the client,
-    though we should log them for our own purposes.
-    """
-    status_code = 500
 
 
 class ValidationError(APIError):
@@ -242,7 +215,7 @@ def request_handler(request):
     except APIError as e:
         logger.debug('Invalid call to api function %s, raised exception: %r',
                      f.__name__, e)
-        return e.response()
+        return Response(e.response_body(), status=e.status_code)
     except ServerError as e:
         logger.error('Server-side failure in function %s, raised exception: %r',
                      f.__name__, e)
